api/fastapi/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sse-starlette>=1.6.0
pydantic>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
dashscope>=1.0.0
redis>=4.5.0
replicate>=0.25.0
ddtrace>=2.0.0
datadog>=0.47.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
chromadb>=0.4.0
numpy>=1.21.0
